graphql-codegen * Rename realtimeVehcilesService to apollo-client and moved it to /src
