+++ src/logic/get-stop-places-with-lines/getStopPlacesWithLines.ts
@@ -1,7 +1,7 @@
 import { TransportMode, TransportSubmode } from '@entur/sdk'
 import { Line, StopPlaceWithLines } from '../../types'
 import { isNotNullOrUndefined, unique } from '../../utils'
-import { apolloClient } from '../../services/realtimeVehicles/realtimeVehiclesService'
+import { apolloClient } from '../../apollo-client'
 import STOP_PLACES_QUERY from './GetStopPlacesWithLines.journey-planner.graphql'
 
 type GetStopPlacesVariables = {
