+++ src/logic/use-stop-places-with-departures/useStopPlacesWithDepartures.ts
@@ -6,7 +6,7 @@ import { isNotNullOrUndefined, nonEmpty, unique } from '../../utils'
 import { useSettings } from '../../settings/SettingsProvider'
 import { REFRESH_INTERVAL } from '../../constants'
 import { useNearestPlaces } from '../use-nearest-places/useNearestPlaces'
-import { apolloClient } from '../../services/realtimeVehicles/realtimeVehiclesService'
+import { apolloClient } from '../../apollo-client'
 import GET_STOP_PLACES_WITH_DEPARTURES_QUERY from './GetStopPlacesWithDepartures.journey-planner.graphql'
 
 type GetStopPlacesWithDeparturesVariables = {
